automation-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
twilio==8.10.3
sendgrid==6.10.0